+++ e2e/actions/actions.go
@@ -2592,5 +2592,6 @@ func E2ETests(env e2e.TestEnv) testhelper.Tests {
 		"ociCdi":     c.actionOciCdi,        // singularity exec --oci --cdi
 		"ociIDMaps":  c.actionOciIDMaps,     // check uid/gid mapping on host for --oci as user / --fakeroot
 		"ociCompat":  np(c.actionOciCompat), // --oci equivalence to native mode --compat
+		"ociOverlay": (c.actionOciOverlay),  // --overlay in OCI mode
 	}
 }